[Feature] Lolipop Object Functionality
Implemented an Assembly Object script which should be attached to a prefab object that will then be spawned into the scene.
Assemby Objects have a Properties value which is a struct containing a Color value and a Quaternion value.

Objects affected:
I have overwritten the TestModule and ModuleAssemblyController scripts to function with the AssemblyObject instead of AssemblyLineObject.
Any place that used the AssemblyLineObject should be replaced with AssemblyObject instead for expected functionality.

Usage:
Assembly Object Scene and connect the lines as usual. Using the Spawn button, objects should now move out properly. Comparison has been tested to work.
Can definitely be optimized in terms of property creation, but this was intended for the Prototype in order to move things along.

Bug report:
Connecting the lines in a travel path causes the objects to travel completely randomly and unpredictably.

Bug Fix:
Fixed the bug stated above by using Array.Reverse. Provided by Toto.