[Feature] - AssemblyObject Lolipop.
Major rehaul of the properties script.
An object will now only have the properties that are actually assigned to it, and not every property possible set.
This is effectively done via a dictionary in the Properties script.

Properties are now classes and not structs, a new property can't be made directly from the editor.

AssemblyObject now have a scriptableObject reference AssetProperties, which is where the properties of each asset are set. On run these are converted to Properties for the Assembly Object.