[Feature] - Else Gate Module
You can now place a Else Gate Module prefab into a scene and select which property it should compare for. If the AssemblyObject has a property with the same value, it will take output path 0, otherwise it will take output path 1.

Modules are a type that are being worked on by multiple people, I've changed the testmodule to work via inheriting from a BaseModule which retains important functionality for travel which is being refactored currently.
ElseGateModule also inherits from this BaseModule. Future Modules should try to adhere to this format.

Usage:
Enter the scene "Else Gate Module" notice the two new modules placed in the scene. One compares for a color, the other for a rotation. I've changed the rotation of the lolipop SO to test this functionality.
Drag a line from the left test Module into the input of the Else Gates, and adjust the property parameters as see fit. When run AssemblyObjects with that property should adhere to the correct path.

Objects Modified:
Mainly modified testModule and it's script functionality, yet should function as expected in other scenes.

#862khkwc0[Review]