[BugFixing] - Multiple Properties Comparison Fixed
Bug identified was caused by another internal bug. Properties used to use Quaternions for rotation, which they still do in the logic, but setting the rotation is now done via a Vector3, that is then converted to a Quaternion when requested.
Bug occured when the ScriptableObject was changed in rotation, where it's Quaternion values would be set despite trying to reset values in inspector.

Removed Debugging logs related to the problem.
#862khxbtd[Review]