[Feature] - Place Modules
Modules can now be placed into a scene given the following:

There is a GameObject with the Player script.
There is a GameObject with the ModuleManager Script - There are prefabs available for both in Prefabs -> Modules

Usage
When in a scene with the prerequsites above, press B to enter build mode, this allows you to place a module onto the ground plane. A transparent sphere is present in scene to represent the can build bool.
Additionally you can also see in Scene view, with Gizmos on the sphere wireframe that's actually being checked.
Press E for a Else Gate Module.
Press R for a Turn Module.
Press B again to exit build mode. Drag and drop path as regular in BuildMachineState.

Currently you cannot change the properties or values of the modules in game, and we are instantiating prefabs. This would be the next step.

Also I went to work with Unity's newer input system. and have setup a PlayerControls under the Player folder in scripts. Look at the Player script for an example of how to integrate the system with scripting.

#862khx863[Review]