Category: [BugFix] SOS
Title - Short explanation:
Changed so 'GameEvent' subscriptions now always work, also after resetting the scene. Changed 'Listening' and 'Not Listening' to 'Subscribe on Play' and 'Do not Subscribe' so better communicate their purpose. Added an 'Invoke' button to the ScriptableGameEvent.

Objects Affected:
All refs should work as normal.

Usage:
It is no longer possible to not listen to a subscribed GameEvent, it is instead possible to choose 'Do not Subscribe' to set a specific ref to not allow subscribtions through it, on play.

Git-ClickUp:
#862khz50w[Review],#862khz51c[Review]