[Feature] Removing modules and replaying levels
Added function for removing modules to the ModuleBase class
The ModuleBase class also has a ModuleIsRemoveable bool that defaults to true. It is currently set to false in the container and spawner modules. If the bool is false, the module cannot be removed by the remove function

Also added buttons for restarting a level (on a win/loss) and retrying from the build phase with the current setup (on a loss) to the GameUI prefab

#862kj4bdw[Review]