[Feature] - Properties Refactoring to predetermined Values.
----- DEPRECATED -----
AssetProperties is deprecated and should be replaced with ObjectProperties - Will delete in following clean-up commit.

Objects that should have properties now use the ObjectProperties scriptable object instead, AssemblyObject has been updated to reflect this.

A scene should now contain a object with the LevelPropertiesHolder script.
A level can now have specific properties via the LevelProperties scriptable object. This will allow objects created, and modules that set properties in one way or another to use a drop down list of possible entries rather than having to set all the values everywhere.

Objects affected:

Everything that used AssetProperties before.

Usage:

Create a new LevelProperty object, and set the values as seen fit. The string Name will be what we reference the values by when looking at other parts of the editor.

New ObjectProperties need a reference to the LevelProperty and can then set the values as defined in the LevelProperty. Upon start these are then converted into a Properties.

ContainerModule now needs a reference to LevelProperty in order to create the list of items it needs to compare against.

#862kj4e0q[Review]