[Feature] - Merge Module
A quick merge module I setup to increase the possible complexity of our levels given the levels.

Takes no consideration as to when something arrives to module. Probably just stacks it if the input arrives at the same time.

Usage:
When in build mode. Press T to insert a merge module.
The module has two inputs and one output. Doesn't do anything extra.