Category: [Feature] Decentralization of LevelProperties and data
Title - Short explanation:
Added a 'TryGetSceneByName' to the LevelData class, and added debug messages to LevelDataHolder class.

Objects Affected:
No Object affected.

Usage:
Use the 'TryGetSceneByName' to attempt to get a scene based on the name - the button will only pop-up once the input name is different than the saved.