Category: [Feature] Decentralization of LevelProperties and data
Title - Short explanation:
Implemented LevelDataGetter everywhere LevelProperties were used, and marked LevelPropertiesHoldler as decrepit. I also added all level scenes to the build settings, and in the LevelDataHolder SO.

Objects Affected:
ContainerModule
ElseGateModule
SpawnerModule
ObjectProperties
Properties

Usage:
It is only possible to set scene by index / name, while in that specific scene is loaded, and only if the scene has been added in Build Settings.
Remember to add new scenes to the build index - this is needed for now.