[Feature] Spawners will now spawn ghost objects in the build phase
Ghost objects behave like normal objects but do not give any points. They can be used to preview a lollipops path through the factory while in the build phase.
Spawners will only spawn one object from a wave before moving on to the next wave instead of spawning the full wave. When it reaches the last wave, it will loop around to the first wave again.

Known problems:
Currently ghost objects are not removed when the run phase starts, but they probably should be.
The visuals of ghost objects and normal objects are currently the same.

Objects Affected:
Spawner module prefab

#862kjuj4q