[Feature] - Unity tests are now available for setup and deployment.
I have sorted and juggled the project a bit around, in order to get the Assembly Definitions correctly working.
SOS has been split into a plugin and a SOS_Extra folder in our scripts, SOS_Extra should be for everything project specific. And SOS itself remains generic and can be further developed there for a package or something.
Placement of reference ScriptableObjects is open for debate.

Because this has been set into an assembly definition I forsee in the future that there will be an issue if you're trying to access something from a package, or unity system (Such as Unity.AI) then it won't be able to find the namespace.
To this end you should find the MainAssembly.asmdef in the Scripts folder, and ensure the assembly definition reference list contains a reference to the appropriate package. Examples in the file should already be available.

Usage:
In Scripts/Tests/EditModeTests, or Scripts/Tests/PlayModeTests you can now write unit tests which can be run and checked using the Unity Test Runner. (Window -> General -> Test Runner)
Write Feature tests, a ModulesFeature should contain all the tests relevant to modules, from placement, to appropriate action upon entry of objects, to correctly accessing Level properties.
Test specific. Rather create 20 tests per feature, as long as the tests themselves are relatively simple. Play tests can generally be a bit more complicated.