[Review] Placement Module, Module Manager Selection tests.
First set of reviews are in, files should be following wiki formatting lines and so forth.

Discovery: Writing Unity tests can be a tedious process that definitely takes time.
Having test writing part of review is quite exhaustive. Tests should start being written in feature development, and the tests should then be reviewed as part of the reviewing process.
Added the GameEvents to ModulesManager: ModulePlacedEvent, and ModuleSelectedEvent.

Usage:
Open the Test Runner Window, and run the Play Mode ModulesFeatureTests.

Expand the ModulesFeatureTests with tests relevant to modules as see fit.

#862khx863[Done]
#862kjh1w3
#862kjh1u2