[Feature] Added events to both remove and place modules based on current understanding.
As far as I have understood the click-up local events refer to a private Action<T> where T is the described relevant object. These events require a subscribe/unsubscribe functionality which has been implemented for public use.

This is likely an incorrect understanding and will probably be reevaluated or changed upon feedback.