Category: [Feature] Visual Prototype 01
Title - Short explanation:
Implemented a new visual prototype with updated assets and playfield. This scene is not yet to be used, and the same for the assets in it, as they are purely visual.

Objects Affected:
All prefabs found inside Prefabs->Greyboxing->VisualPrototype.
CineMachine has been added as package.
Feel has been added as package.

Usage:
This scene is not yet to be used, and the same for the assets in it, as they are purely visual.

#862kkz06t[Review]