[Design] - A UML diagram for the Grid System.
The Diagram is based on the design documentation Toto provided earlier, with some aspects of the code that Andreas had already contributed towards.

To use or view the diagram. For VSCode install the plugin PlantUML, For Vstudio find the plugin on the microsoft marketplace also named PlantUML.
In both cases you'll need to (for now) install a local PlantUML java program so that it can process and display the diagrams. The alternative is setting up a PlantUML server which we might look into in the future.

Once installed, all you have to do to view a diagram, is 1. preview current diagram, by going into a supported extension file, (primarily use .puml) right click in the text editor and select preview current diagram. This diagram is real-time updating meaning you can visualize your changes after a very brief update.

Additionally diagrams can be exported into PNGs although I prefer the preview due to it being vector images, and pngs are limited by size.