[Design] - Worked on more design regarding grid.
Added descriptions and some implementations as I went on with this.

I decided to stop at this point now, because I am beginning to realize the necessity of an edit tool instead. Which will take some time to make but should provide a more intuitive interaction with the grid from editor mode. Potentially this could also be used as a beginning tool to allowing players to create their own maps.