[Feature] Added BFS pathfinding to the Grid
Description:
implemented BFS for pathfinding in the grid. This should be changed to A* later
Refactored the grid to no longer expose the Cell class. Cell is now only used internally in the grid.
Everything external now uses Vector2Int as a representation of the grid cell coordinate to interact with the grid

#862kmkkf1