[Feature] - Continuous progress on AssemblyLine system.
Last commit before break kind of a status of sorts. System is not fully implemented yet, and has yet to be tested.
Currently key elements missing are recursion in regards to connecting lines, and removal of individual pieces/lines.
System also has a current UML in AssemblySystem.puml. Will definitely be worked more on and expanded but I'm feeling good about this.