[Feature] - Assemblylines can now transport items in a single direction.
This is still not integrated with modules or anything else, and will not work by trying to connect to modules.

A lot of the foundation of multiple connecting assembly lines is there, but quite a lot of the logic is missing.

Currently I've interpreted the assembly lines as a tree structure, where the root line is the assembly line with connecting assembly lines.
there is no protection against loops yet.
If trying to create connecting assembly lines, it will run, but incorrectly as there's no logic for how the connecting lines should work and where the starts of these lines exists.

Next priority is fixing the combining logic of these lines. including removal of pieces.
after that combining the spawner module logic with the lines as to spawn the traveling items correctly rather than spawning test lolipops.