[Refactor, Feature] Slowly moving grid placement logic out of modulesManager and started working on new system for modules
Description:
The logic for checking which grid position is being moused over has been copied over to the MouseInput class. The goal is to get rid of the old modulesManager. It is currently only used to check if we are in build mode
I also started working on a system for defining modules as scriptable objects, but will need to integrate this with the system for assembly lines