[Feature] It is now possible to have gridobjects that occupy more than one cell
Description:
The size of each gridObject is defined in the GridObjectSO scriptable object
Implemented a ModulePlacer that can place modules of any size on the grid. Also adjusted the visual in GridVisual to match
Module selection is now done in the BuildingSelector class. This class has an event for when a new building is selected
All of this is implemented in the new scene ModulePlacementOnGrid. Here the buildings can be selected by pressing the number keys on the keyboard, and can be placed by pressing the right mouse button