[Refactor, Feature] - AssemblyLineSystem got substantially big, moved specific AssemblyLineManagement logic to AssemblyLineManager.
One should still only have to interact with the AssemblyLineSystem.

Mostly implemented loops, I keep discovering more potential edge cases when it comes to line merging. It is definitely not a completed product yet, I think connections are lining up correctly now, but frankly I haven't tested every possible combination.
Loops should now also be possible.

Large amounts of rework and tweaking will probably be introduced once removal of assembly lines becomes the next thing to work on.

Combining GridObjectSO and TestingScript:
Good suggestion of making a CreateInstance function that can be overridden by SO that extend GridObjectSO.
Rotation of the silihouette object. Please try and move it to mousewheel or something so we don't have to remember every keybinding.