[Feature] - Debug visualization added to assembly lines
The AssemblyLineSystem needs a AssemblyLineVisualizer script which needs a reference to grid.

An assembly line will be represented as A circle at a start location, and Triangle at an end position.
Connecting lines of an assembly lines will have smaller circle/triangle of that line color.

A new assembly line should receive a unique color. (upper limit of course)