[Feature] - Initial deployment of Interface for transporting items.
I have currently suggested an implementation of two variations of transportable units a ITransportable and an IStorageTransportable.

The difference between the two is that the storage should be able to hold multiple items to a storage capacity before being Occupied.
Refactored PieceState to be TransportState located in the ITransportable interface.