[Refactor] - Removed redundant linking system from AssemblyLine, AssemblyPiece, and ITransportable.
Generally most link behavior, about where to send pieces and such should not be available via the AssemblyLine.
ITransportable now includes a function that sets the parentAssemblyLine alongside the node in the linked list, the node is purely for optimization purposes.
AssemblyLine and Manager now functions based on the linked list, and looping behavior has been updated to work with this. Along with some Recursion safeguards when debugging. Removed some unused parts of IGridObject.