[Refactor] Major refactor for the grid system
Description:
The grid system now has multiple different classes instead of only having a grid class. The main classes are
CellGrid: Generic class that allows placement of object in a grid. This is not a monobehaviour
GridLayout: Converts between grid space and world space. Contains information about grid rotation, origin and cell size
FactoryGrid: A wrapper for the two above that will be the main grid that will be used in the game. Therefore this should be considered part of the game and not part of the grid system
GridCellSpawner: Spawns a prefab on each cell position in a grid