[Feature] Implemented spawner module and a simple editor for the grid
Description:
The testModule2x2 now acts as a spawner module
It is now possible  to "Paint" on the GridCellSpawner to change where it should spawn grid cells. If a grid position does not have a prefab cell, it is considered occupied for building and pathfinding purposes.
In practice this means that you cannot build on positions that do not have a "ground" prefab