[Bugfix] - Ton of connection bugs regarding tunnels should be fixed.
Description:
Most connection related bugs when trying to connect multiple tunnels together or removing them should be fixed.

1 persisting issue is that if you place a tunnel, and two assembly pieces on either side of the tunnel, then it will directly link these pieces together.
However, if you then afterwards place a tunnel that would connect to the original tunnel then that connection is established and the whole thing is a mess.

Needs a way of locking receiving or connecting if either side of the tunnel is connected to an assembly piece.

Currently best practice for testing and making maps is to place tunnels first, then connect them with assembly pieces after.